adding geopoint support
