+++ hat/assets/js/apps/Iaso/domains/entities/details.tsx
@@ -1,26 +1,26 @@
-import React, { FunctionComponent, useMemo } from 'react';
-import {
-    useSafeIntl,
-    commonStyles,
-    useGoBack,
-    LoadingSpinner,
-    LinkButton,
-} from 'bluesquare-components';
 import { Box, Divider, Grid } from '@mui/material';
 import { makeStyles } from '@mui/styles';
-import TopBar from '../../components/nav/TopBarComponent';
-import MESSAGES from './messages';
-import { baseUrls } from '../../constants/urls';
-import { useGetBeneficiary, useGetSubmissions } from './hooks/requests';
-import { useGetBeneficiaryFields } from './hooks/useGetBeneficiaryFields';
-import { Beneficiary } from './types/beneficiary';
-import { useBeneficiariesDetailsColumns } from './config';
+import {
+    commonStyles,
+    LinkButton,
+    LoadingSpinner,
+    useGoBack,
+    useSafeIntl,
+} from 'bluesquare-components';
+import React, { FunctionComponent, useMemo } from 'react';
 import { CsvButton } from '../../components/Buttons/CsvButton';
 import { XlsxButton } from '../../components/Buttons/XslxButton';
-import { BeneficiaryBaseInfo } from './components/BeneficiaryBaseInfo';
+import TopBar from '../../components/nav/TopBarComponent';
 import WidgetPaper from '../../components/papers/WidgetPaperComponent';
 import { TableWithDeepLink } from '../../components/tables/TableWithDeepLink';
+import { baseUrls } from '../../constants/urls';
 import { useParamsObject } from '../../routing/hooks/useParamsObject';
+import { BeneficiaryBaseInfo } from './components/BeneficiaryBaseInfo';
+import { useBeneficiariesDetailsColumns } from './config';
+import { useGetBeneficiary, useGetSubmissions } from './hooks/requests';
+import { useGetBeneficiaryFields } from './hooks/useGetBeneficiaryFields';
+import MESSAGES from './messages';
+import { Beneficiary } from './types/beneficiary';
 
 const useStyles = makeStyles(theme => ({
     ...commonStyles(theme),
@@ -73,17 +73,58 @@ export const Details: FunctionComponent = () => {
             {!isLoading && (
                 <Box className={`${classes.containerFullHeightNoTabPadded}`}>
                     <Grid container spacing={2}>
-                        <Grid container item xs={4}>
+                        <Grid container item xs={12} md={4}>
                             <BeneficiaryBaseInfo
                                 beneficiary={beneficiary}
                                 fields={beneficiaryFields}
                             />
                         </Grid>
+
+                        <Grid container item xs={12} md={8}>
+                            <WidgetPaper
+                                className={classes.fullWidth}
+                                title={formatMessage(MESSAGES.submissions)}
+                            >
+                                <TableWithDeepLink
+                                    marginTop={false}
+                                    countOnTop={false}
+                                    elevation={0}
+                                    baseUrl={baseUrls.entityDetails}
+                                    data={data?.instances ?? []}
+                                    pages={data?.pages}
+                                    defaultSorted={[{ id: 'id', desc: false }]}
+                                    columns={columns}
+                                    count={data?.count}
+                                    params={params}
+                                    extraProps={{
+                                        loading: isLoadingSubmissions,
+                                    }}
+                                />
+                                <Divider />
+                                <Box display="flex" py={2}>
+                                    <Box mr={1} ml={2}>
+                                        <CsvButton
+                                            csvUrl={`/api/entities/?csv=true&id=${entityId}`}
+                                        />
+                                    </Box>
+                                    <XlsxButton
+                                        xlsxUrl={`/api/entities/?xlsx=true&id=${entityId}`}
+                                    />
+                                </Box>
+                            </WidgetPaper>
+                        </Grid>
+                        {/* <Grid container item xs={7} justifyContent="flex-end">
+                            <Box>
+                                <LinkButton to={duplicateUrl}>
+                                    {formatMessage(MESSAGES.seeDuplicates)}
+                                </LinkButton>
+                            </Box>
+                        </Grid> */}
                         {duplicates.length > 0 && (
                             <Grid
                                 container
                                 item
-                                xs={8}
+                                xs={7}
                                 justifyContent="flex-end"
                             >
                                 <Box>
@@ -93,57 +134,7 @@ export const Details: FunctionComponent = () => {
                                 </Box>
                             </Grid>
                         )}
-                        {/* TODO uncomment when edition is possible */}
-                        {/* <Grid container item xs={1}>
-                        <Box ml={2}>
-                            <EditIcon
-                                onClick={() => {
-                                    console.log(
-                                        'Edit Beneficiary',
-                                        beneficiary?.name,
-                                        entityId,
-                                    );
-                                    // eslint-disable-next-line no-alert
-                                    alert('Entity edition');
-                                }}
-                                color="action"
-                            />
-                        </Box>
-                    </Grid> */}
                     </Grid>
-                    <Box mt={2}>
-                        <WidgetPaper
-                            className={classes.fullWidth}
-                            title={formatMessage(MESSAGES.submissions)}
-                        >
-                            <TableWithDeepLink
-                                marginTop={false}
-                                countOnTop={false}
-                                elevation={0}
-                                baseUrl={baseUrls.entityDetails}
-                                data={data?.instances ?? []}
-                                pages={data?.pages}
-                                defaultSorted={[{ id: 'id', desc: false }]}
-                                columns={columns}
-                                count={data?.count}
-                                params={params}
-                                extraProps={{
-                                    loading: isLoadingSubmissions,
-                                }}
-                            />
-                            <Divider />
-                            <Box display="flex" py={2}>
-                                <Box mr={1} ml={2}>
-                                    <CsvButton
-                                        csvUrl={`/api/entities/?csv=true&id=${entityId}`}
-                                    />
-                                </Box>
-                                <XlsxButton
-                                    xlsxUrl={`/api/entities/?xlsx=true&id=${entityId}`}
-                                />
-                            </Box>
-                        </WidgetPaper>
-                    </Box>
                 </Box>
             )}
         </>
